NEW support for MySQL connections with SSL certificates
