+++ Adminer/adminer.php
@@ -2,7 +2,7 @@
 
 function adminer_object() {
     global $adminer;
-    
+	    
     // required to run any plugin
     require_once "../plugins/plugin.php";
     
@@ -11,9 +11,9 @@ function adminer_object() {
         require_once $filename;
     }
 
-	return new AdminerPlugin([
+	$plugins =[
 	    new AdminerDesign('exface'),
-		// TODO: inline the result of password_hash() so that the password is not visible in source codes
+		// TODO use constant AdminerAPI::NO_PASSWROD
 	    new AdminerLoginPasswordLess(password_hash(12345678, PASSWORD_DEFAULT)),
 		new AdminerTablesFilter(),
 		new AdminerForeignKeys(),
@@ -23,7 +23,23 @@ function adminer_object() {
 	    new AdminerSaveMenuPos(),
 	    new AdminerTreeViewer('plugins/tree-viewer/script.js'),
 	    new AdminerFrames(true),
-	]);
+	];
+
+	// See if the AdminerAPI requires SSL for this connection.
+	// If so, make sure to remember this setting from the login call and
+	// remember it till the next login attempt.
+	if (array_key_exists('auth', $_POST)) {
+		if (null !== $ssl = $_POST['auth']['ssl'] ?? null) {
+			$_SESSION['ssl'] = $_POST['auth']['ssl'];
+		} else {
+			unset($_SESSION['ssl']);
+		}
+	}
+	if (array_key_exists('ssl', $_SESSION)) {
+		$plugins[] = new AdminerLoginSsl($_SESSION['ssl']);
+	}
+	
+	return new AdminerPlugin($plugins);
 }
 $dir = __DIR__;
 chdir($dir . "/adminer/");
